Improve beacon scheduling + fix GPS epoch nsec typo.
Instead of converting the GPS epoch time to the counter_us value for
scheduling, the OnGPS scheduling type is used to transmit the beacon on
the next PPS. This also compensates the tx frequency with the xtal
correction value.

The count_us value is still needed so that the JIT queue is triggered on
the right time.