Remove Class-B beacon frequency correction.
This was copied over from the Semtech UDP Packet Forwarder, but should
have been implemented on dequeue instead of enqueue, as the enqueue is
happening in the future. For now this is removed, as it might need
further research if this logic is needed or not. It is present in the
Semtech UDP Packet Forwarder, but I could not find the same logic in the
Basics Station.

If this is needed, it must be implemented on dequeue, to make sure that
we compensate the frequency with the xtal_correct value as close to the
actual transmission as possible.

For now this is removed as on sending the downlink, the Concentratord
aggregates the downlink count on frequency. Before this change it would
generate a many different frequencies as the xtal_correct value is
variable.