Update UBX binary format parsing.
As the UBX binary format indicates the payload length, we can
immediately read the full payload. This avoids errors in case the UBX
message does not start with B5620120 and the payload contains a byte
equal to the ASCII $ character, in which case the remaining UBX payload
will be parsed as NMEA and the parsing returns an error.