Refactor ZMQ interfaces (events and commands).
This refactors the ZMQ interface from multipart messages to single
messages.

Previously, the first ZMQ message would contain a string indicating
the payload (e.g. up, stats, ...) and the second message would contain
the Protobuf payload. This is prone to errors, as the string indicating
the payload can not be type checked.

This refactors the interface to use a single ZMQ payload, which can
immediately be encoded / decoded using Protobuf using the Event or
Command message. As in Rust the event of the Event message and command
field of the Command message are implemented as enum, the compiler is
able to check that all enum variants are matched.

Users implementing against the Concentratord ZMQ interface must make
some small changes, in their ZMQ receiving and sending logic as this
refactor is not backwards compatible.